asteroid_data/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
boto3==1.28.3
botocore==1.31.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-datastore==2.16.1
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
yarl==1.9.2
